backend/requirements.txt
fastapi
uvicorn
pydantic
sqlalchemy
databases
chromadb
ollama
pyyaml
